requirements.txt
absl-py==2.1.0
astroid==3.3.5
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.0
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.1
Pillow==9.3.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylint==3.3.1
pytest==8.3.3
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-intel==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
Werkzeug==3.0.6
wrapt==1.16.0
